Clean up the farm code to better isolate functionality, and make the comps more, well, composable
