+++ 1.5/Defs/Buildings/Buildings_Structure.xml
@@ -152,7 +152,7 @@
                 <colonyAnimalsOnly>true</colonyAnimalsOnly>
                 <showGizmo>true</showGizmo>
             </li>
-            <li Class="PawnStorages.Farm.CompProperties_FarmStorage">
+            <li Class="PawnStorages.Farm.Comps.CompProperties_FarmStorage">
                 <needsDrop>true</needsDrop>
                 <releaseOption>true</releaseOption>
                 <convertOption>false</convertOption>
@@ -160,10 +160,9 @@
                 <lightEffect>true</lightEffect>
                 <storageStation>PS_BatteryFarm</storageStation>
             </li>
-            <li Class="PawnStorages.Farm.CompProperties_FarmNutrition">
-                <maxNutrition>500</maxNutrition>
-                <doesProduction>true</doesProduction>
-                <doesBreeding>false</doesBreeding>
+            <li Class="PawnStorages.Farm.Comps.CompProperties_FarmNutrition">
+            </li>
+            <li Class="PawnStorages.Farm.Comps.CompProperties_FarmProducer">
             </li>
         </comps>
         <building>
@@ -249,7 +248,7 @@
                 <colonyAnimalsOnly>true</colonyAnimalsOnly>
                 <showGizmo>true</showGizmo>
             </li>
-            <li Class="PawnStorages.Farm.CompProperties_FarmStorage">
+            <li Class="PawnStorages.Farm.Comps.CompProperties_FarmStorage">
                 <needsDrop>true</needsDrop>
                 <releaseOption>true</releaseOption>
                 <convertOption>false</convertOption>
@@ -257,10 +256,9 @@
                 <lightEffect>true</lightEffect>
                 <storageStation>PS_BatteryFarm</storageStation>
             </li>
-            <li Class="PawnStorages.Farm.CompProperties_FarmNutrition">
-                <maxNutrition>500</maxNutrition>
-                <doesProduction>false</doesProduction>
-                <doesBreeding>true</doesBreeding>
+            <li Class="PawnStorages.Farm.Comps.CompProperties_FarmNutrition">
+            </li>
+            <li Class="PawnStorages.Farm.Comps.CompProperties_FarmBreeder">
             </li>
         </comps>
         <building>
