+++ 1.5/Source/PawnStorages/PawnStorages/Farm/Comps/CompProperties_FarmBreeder.cs
@@ -1,0 +1,11 @@
+﻿using Verse;
+
+namespace PawnStorages.Farm.Comps;
+
+public class CompProperties_FarmBreeder : CompProperties
+{
+    public CompProperties_FarmBreeder()
+    {
+        compClass = typeof(CompFarmBreeder);
+    }
+}
