+++ 1.5/Source/PawnStorages/PawnStorages/Farm/Comps/CompProperties_FarmProducer.cs
@@ -1,0 +1,11 @@
+﻿using Verse;
+
+namespace PawnStorages.Farm.Comps;
+
+public class CompProperties_FarmProducer : CompProperties
+{
+    public CompProperties_FarmProducer()
+    {
+        compClass = typeof(CompFarmProducer);
+    }
+}
