+++ 1.5/Source/PawnStorages/PawnStorages/Farm/Interfaces/IActive.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PawnStorages.Farm.Interfaces
+{
+    public interface IActive
+    {
+        public bool IsActive { get; }
+    }
+}
