+++ 1.5/Source/PawnStorages/PawnStorages/Farm/Interfaces/IBreederParent.cs
@@ -1,0 +1,18 @@
+﻿using PawnStorages.Farm.Interfaces;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Verse;
+
+namespace PawnStorages.Farm.Interfaces 
+{
+    public interface IBreederParent : IActive
+    {
+        public List<Pawn> BreedablePawns { get; }
+        public int TickInterval { get; }
+
+        public void StoreNewPawn(Pawn newPawn);
+    }
+}
