+++ 1.5/Source/PawnStorages/PawnStorages/Farm/Interfaces/INutritionStorageParent.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using PawnStorages.Farm.Interfaces;
+using Verse;
+
+namespace PawnStorages.Farm.Interfaces
+{
+    public interface INutritionStorageParent : IActive, IPawnRelease
+    {
+        public bool HasSuggestiveSilos { get; }
+        public bool HasStoredPawns { get; }
+        public List<Pawn> StoredPawns { get; }
+    }
+}
