+++ 1.5/Source/PawnStorages/PawnStorages/Farm/Interfaces/IPawnRelease.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Verse;
+
+namespace PawnStorages.Farm.Interfaces
+{
+    public interface IPawnRelease
+    {
+        public void ReleasePawn(Pawn pawn);
+    }
+}
