+++ 1.5/Source/PawnStorages/PawnStorages/Farm/Interfaces/IProductionParent.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using Verse;
+
+namespace PawnStorages.Farm.Interfaces
+{
+    public interface IProductionParent: IActive, IPawnRelease
+    {
+        public List<Pawn> ProducingPawns { get; }
+        public int TickInterval { get; }
+    }
+}
