+++ 1.5/Source/PawnStorages/PawnStorages/PawnStorages.csproj
@@ -48,7 +48,7 @@
         </Reference>
     </ItemGroup>
 
-    <ItemGroup Condition="'@(Reference->WithMetadataValue('Identity', 'Assembly-CSharp')->Metadata('HintPath')->Exists())' == ''">
+    <ItemGroup Condition="'@(Reference-&gt;WithMetadataValue('Identity', 'Assembly-CSharp')-&gt;Metadata('HintPath')-&gt;Exists())' == ''">
         <PackageReference Include="Krafs.Rimworld.Ref" Version="1.5-*" />
     </ItemGroup>
     <ItemGroup>
